requirements.txt
# Bondlink Server - Production Dependencies
# VPS Server for Multi-Client Tunnel Aggregation

# Network and system monitoring
psutil>=5.9.8
netifaces>=0.11.0
pyroute2>=0.7.12

# Async networking
asyncio>=3.4.3
aiohttp>=3.9.0
websockets>=12.0

# Web UI and API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
bcrypt>=4.1.2

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Database (for client management and stats)
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Logging and monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Encryption and security
cryptography>=42.0.0
pynacl>=1.5.0

# CLI tools
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0
requests>=2.31.0

# System integration
systemd-python>=235; sys_platform == 'linux'

# Development and testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
